third_party/blink/web_tests/platform/generic/fast/deprecated-flexbox/repaint-scrollbar-expected.txt

Check if changing flexbox's content doesn't change the state of scroll. See this bug for detail: https://bugs.webkit.org/show_bug.cgi?id=15135

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".










meroL ipsum
fooooooooooooooo meroL ipsum









meroL ipsum









meroL ipsum
PASS element.scrollTop is 108
PASS element.scrollLeft is 108
PASS element.scrollTop is 108
PASS element.scrollTop is 108
PASS successfullyParsed is true

TEST COMPLETE
